requirements.txt
fastapi==0.109.1
uvicorn==0.27.0
python-dotenv==1.0.1
pdfplumber==0.10.0
gTTS==2.4.0
requests==2.31.0
httpx==0.27.0
xhtml2pdf==0.2.15
python-multipart==0.0.6
# LiveKit dependencies
livekit>=0.11.0
livekit-api>=0.5.0
opencv-python==4.8.1.78
mediapipe==0.10.7
numpy==1.24.3
jinja2==3.1.2
pillow==10.0.0
email-validator==2.1.1
aiosmtplib==3.0.2
# ATS functionality dependencies
PyPDF2==3.0.1
google-generativeai==0.3.2
# Database dependencies
motor>=3.4,<4.0
certifi>=2024.2.2